moved api calls to its own module
