+++ scripts/main.js
@@ -11,9 +11,6 @@
 //   hubot gatherer clash [@handler] - this command will query the service for two random cards and
 //                                     then compare their converted mana costs to see which is higher
 //
-// Notes:
-//   <optional notes required for the script>
-//
 // Author:
 //   HerrPfister and wickerpopstar
 
@@ -21,6 +18,7 @@ var Q = require('q');
 
 var UrlMap = require('../static/consts').urlMap;
 
+var api = require('./utils/api');
 var utils = require('./utils/utils');
 var GathererFind = require('./commands/gFind');
 var GathererClash = require('./commands/gClash');
@@ -28,9 +26,9 @@ var GathererRandom = require('./commands/gRandom');
 
 module.exports = function (robot) {
   robot.respond(/gatherer\s+clash\s+(@\w+)/i, function(robo) {
-    Q.all([utils.getRandomMultiverseId(robot), utils.getRandomMultiverseId(robot)])
+    Q.all([api.getRandomMultiverseId(robot), api.getRandomMultiverseId(robot)])
       .done(function(multiverseIds) {
-        Q.all([utils.getRandomCard(robot), utils.getRandomCard(robot, multiverseIds[1])])
+        Q.all([api.getRandomCard(robot, multiverseIds[0]), api.getRandomCard(robot, multiverseIds[1])])
           .done(function(cards) {
             var challenger = { name: robo.message.user.name, card: cards[0] };
             var challenged = { name: robo.match[1], card: cards[1] };
@@ -41,9 +39,9 @@ module.exports = function (robot) {
   });
 
   robot.respond(/gatherer\s+random/i, function(robo) {
-    utils.getRandomMultiverseId(robot)
+    api.getRandomMultiverseId(robot)
       .then(function(multiverseId){
-        return utils.getRandomCard(robot, multiverseId);
+        return api.getRandomCard(robot, multiverseId);
       })
       .done(function(card){
         GathererRandom.parseResponse(robo, card);
@@ -51,8 +49,6 @@ module.exports = function (robot) {
   });
 
   robot.respond(/gatherer\s+find\s+(.*)/i, function (robo) {
-
-    // Grab the captured user's input and parse for search params
     var cardName = utils.getCardName(robo.match[1]);
     var urlParams = utils.parseUrlParams(robo.match[1]);
